+++ client/components/CardWithIcon/styles.js
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
 
 export const Container = styled.div`
+  width: 1.5rem;
   &:hover {
     cursor: pointer;
   }
