+++ client/components/ImageModal/index.js
@@ -1,13 +1,9 @@
-import Image from 'next/image';
-
 import { Container } from './styles';
 
-export function ImageModal({ closeFunction, photo }) {
+export function ImageModal({ closeFunction }) {
   return (
     <Container>
       <button onClick={closeFunction}>X</button>
-
-      <Image src={photo} alt="An image from a historical moment" />
     </Container>
   );
 }
