+++ client/pages/_document.js
@@ -5,7 +5,10 @@ export default function Document() {
     <Html lang="en">
       <Head>
         <meta charSet="UTF-8" />
-        <meta name="description" content="Generated by create next app" />
+        <meta
+          name="description"
+          content="Computer Science bachelor who loves writing code and teaching code. Worked on big projects and interacted with people from many places"
+        />
         <link rel="icon" href="/favicon.svg" />
         <link rel="preconnect" href="https://fonts.googleapis.com" />
         <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="true" />
