+++ client/styles/index.js
@@ -6,7 +6,6 @@ export const Container = styled.div`
   background-repeat: no-repeat;
   background-attachment: fixed;
   background-size: cover;
-
   width: 100%;
 `;
 
@@ -14,7 +13,6 @@ export const Main = styled.main`
   width: 90%;
   height: 100vh;
   margin: 0 auto;
-
   display: flex;
   align-items: center;
   justify-content: center;
@@ -30,7 +28,6 @@ export const Title = styled.h1`
   font-family: ${({ theme }) => theme.FONTS.MAIN};
   font-weight: 300;
   font-size: 1.2rem;
-
   padding: 1rem 0;
 `;
